# Optimized _beans_when_has_action_do_render()