@hellofromtonya 2018-03-21
BTW @iCaspar and @christophherr  - This is an example of using the [Micro Profiler](https://github.com/hellofromtonya/beans-profiler) to measure the performance of each function as we do our work. This tool is not ready yet for general usage.  But it's a handy tool for validating that we do not have negative impacts to Beans' execution time.